requirements.txt
matplotlib==3.7.1
numpy==1.24.2
Pillow==9.5.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
kiwisolver==1.4.4
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
keyboard~=0.13.5
selenium~=4.8.3
